Allow not only english language file uploads
Currently a student as well as staff member could
upload any file as part of submission or grade.

Unfortunately when the user trying to DOWNLOAD
the files that filename is not in English (Russian for example)
we got an 500 error on the server.

This occurs because of Python encoding rules, and original
issue is something like this:

"UnicodeEncodeError: 'latin-1' codec can't encode characters
in position 21-24: ordinal not in range(256)"

(Depends on the filename)

This patch is fixing it.